--- lib/forgot_password2/forgot_password2_model.dart
@@ -1,53 +1,0 @@
-import '/flutter_flow/flutter_flow_icon_button.dart';
-import '/flutter_flow/flutter_flow_theme.dart';
-import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
-import 'forgot_password2_widget.dart' show ForgotPassword2Widget;
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:provider/provider.dart';
-
-class ForgotPassword2Model extends FlutterFlowModel<ForgotPassword2Widget> {
-  ///  State fields for stateful widgets in this page.
-
-  final unfocusNode = FocusNode();
-  // State field(s) for txtxCodigo widget.
-  FocusNode? txtxCodigoFocusNode;
-  TextEditingController? txtxCodigoController;
-  String? Function(BuildContext, String?)? txtxCodigoControllerValidator;
-  // State field(s) for password widget.
-  FocusNode? passwordFocusNode;
-  TextEditingController? passwordController;
-  late bool passwordVisibility;
-  String? Function(BuildContext, String?)? passwordControllerValidator;
-  // State field(s) for passwordConfirm widget.
-  FocusNode? passwordConfirmFocusNode;
-  TextEditingController? passwordConfirmController;
-  late bool passwordConfirmVisibility;
-  String? Function(BuildContext, String?)? passwordConfirmControllerValidator;
-
-  /// Initialization and disposal methods.
-
-  void initState(BuildContext context) {
-    passwordVisibility = false;
-    passwordConfirmVisibility = false;
-  }
-
-  void dispose() {
-    unfocusNode.dispose();
-    txtxCodigoFocusNode?.dispose();
-    txtxCodigoController?.dispose();
-
-    passwordFocusNode?.dispose();
-    passwordController?.dispose();
-
-    passwordConfirmFocusNode?.dispose();
-    passwordConfirmController?.dispose();
-  }
-
-  /// Action blocks are added here.
-
-  /// Additional helper methods are added here.
-}
