+++ lib/index.dart
@@ -1,8 +1,6 @@
 // Export pages
 export '/login/login_widget.dart' show LoginWidget;
-export '/inicio/inicio_widget.dart' show InicioWidget;
+export '/home/home_widget.dart' show HomeWidget;
 export '/create_account/create_account_widget.dart' show CreateAccountWidget;
-export '/forgot_password1/forgot_password1_widget.dart'
-    show ForgotPassword1Widget;
-export '/forgot_password2/forgot_password2_widget.dart'
-    show ForgotPassword2Widget;
+export '/forgot_password/forgot_password_widget.dart' show ForgotPasswordWidget;
+export '/password/password_widget.dart' show PasswordWidget;
